ISOLATION CORRECTION V1
