+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,60 +7,20 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.app.Regional_News.EnewspaperActivity$HomeFragment"
+        android:name="com.app.Regional_News.HomeFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/nav_event"
-        android:name="com.app.Regional_News.EnewspaperActivity$SearchFragment"
+        android:name="com.app.Regional_News.SearchFragment"
         android:label="@string/menu_search"
         tools:layout="@layout/fragment_search" />
 
-    <fragment
-        android:id="@+id/nav_visitor"
-        android:name="com.app.Regional_News.ui.slideshow.VisitorFragment"
-        android:label="@string/menu_visitor"
-        tools:layout="@layout/fragment_visitor" />
 
     <fragment
-        android:id="@+id/nav_notification"
-        android:name="com.app.Regional_News.ui.gallery.NotificationFragment"
-        android:label="@string/menu_noti"
-        tools:layout="@layout/fragment_notification" />
-    <fragment
-        android:id="@+id/nav_complaint"
-        android:name="com.app.Regional_News.ui.gallery.ComplainFragment"
-        android:label="@string/menu_complaint"
-        tools:layout="@layout/fragment_complaint" />
-    <fragment
-        android:id="@+id/nav_addcomplaint"
-        android:name="com.app.Regional_News.ui.gallery.AddComplainFragment"
-        android:label="@string/menu_acomplaint"
-        tools:layout="@layout/fragment_addcomplaint" />
-    <fragment
-        android:id="@+id/nav_wing"
-        android:name="com.app.Regional_News.ui.gallery.WingFragment"
-        android:label="@string/menu_wing"
-        tools:layout="@layout/fragment_wing" />
-    <fragment
-        android:id="@+id/nav_staff"
-        android:name="com.app.Regional_News.ui.gallery.StaffFragment"
-        android:label="@string/menu_staff"
-        tools:layout="@layout/fragment_staff" />
-    <fragment
-        android:id="@+id/nav_income"
-        android:name="com.app.Regional_News.ui.gallery.IncomeFragment"
-        android:label="@string/menu_income"
-        tools:layout="@layout/fragment_income" />
-    <fragment
-        android:id="@+id/nav_expense"
-        android:name="com.app.Regional_News.ui.gallery.ExpenseFragment"
-        android:label="@string/menu_expense"
-        tools:layout="@layout/fragment_income" />
-    <fragment
         android:id="@+id/nav_setting"
-        android:name="com.app.Regional_News.EditorialFragment$SettingFragment"
+        android:name="com.app.Regional_News.SettingFragment"
         android:label="@string/menu_setting"
         tools:layout="@layout/fragment_setting" />
